third_party/blink/web_tests/editing/selection/modify_move/move_page_down_inline_block_hang.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div contenteditable="true" style="display: inline-block;" id=target>
<div style="display: inline-block;">test<br><br></div>
</div>
<script>
// Regression test for crbug.com/933839
test(() => {
  assert_own_property(
      window, 'testRunner',
      'This test requires testRunner to perform PageDown selection movement');
  target.focus();
  testRunner.execCommand('MovePageDown');
}, 'PageDown with nested inline blocks does not hang');
</script>
